requirements.txt
aioredis==2.0.1
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
astroid==3.1.0
async-timeout==4.0.3
asyncpg==0.29.0
bcrypt==4.1.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
dill==0.3.8
ecdsa==0.18.0
exceptiongroup==1.2.0
fastapi==0.110.0
greenlet==3.0.3
h11==0.14.0
idna==3.6
isort==5.13.2
Mako==1.3.2
MarkupSafe==2.1.5
mccabe==0.7.0
passlib==1.7.4
platformdirs==4.2.0
pyasn1==0.5.1
pydantic==2.6.2
pydantic_core==2.16.3
pylint==3.1.0
python-dotenv==1.0.1
python-jose==3.3.0
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.27
starlette==0.36.3
tomli==2.0.1
tomlkit==0.12.3
typing_extensions==4.10.0
urllib3==2.2.1
uvicorn==0.27.1
